Fix comments
Delete button is placed in the expanded container.
Use circle to show state of device.
Moved test to separate file.
