+++ frontend/app/src/main/java/dk/scheduling/schedulingfrontend/HomePage.kt
@@ -1,8 +1,10 @@
 package dk.scheduling.schedulingfrontend
 
 import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
@@ -11,14 +13,14 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowDropDown
 import androidx.compose.material.icons.filled.Delete
-import androidx.compose.material.icons.filled.PowerSettingsNew
-import androidx.compose.material.icons.filled.Schedule
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
@@ -35,7 +37,9 @@ import androidx.compose.runtime.setValue
 import androidx.compose.runtime.toMutableStateList
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.rotate
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -43,13 +47,13 @@ import dk.scheduling.schedulingfrontend.device.Device
 import dk.scheduling.schedulingfrontend.device.DeviceOverview
 import dk.scheduling.schedulingfrontend.device.DeviceState
 import dk.scheduling.schedulingfrontend.device.getDeviceState
-import dk.scheduling.schedulingfrontend.device.testDeviceOverview
 import dk.scheduling.schedulingfrontend.sharedcomponents.DATE_FORMATTER
 import dk.scheduling.schedulingfrontend.sharedcomponents.FilledButton
 import dk.scheduling.schedulingfrontend.sharedcomponents.OutlinedButton
 import dk.scheduling.schedulingfrontend.ui.theme.SchedulingFrontendTheme
 import dk.scheduling.schedulingfrontend.ui.theme.scheduled
 import dk.scheduling.schedulingfrontend.ui.theme.success
+import testdata.testDeviceOverview
 import java.time.temporal.ChronoUnit
 
 @Preview(showBackground = true, device = "spec:id=reference_phone,shape=Normal,width=411,height=891,unit=dp,dpi=420")
@@ -140,12 +144,41 @@ fun DeviceCard(
                 )
             }
 
+            Icon(
+                modifier = Modifier.rotate(rotationState),
+                imageVector = Icons.Default.ArrowDropDown,
+                contentDescription = "Drop-Down Arrow",
+            )
+        }
+
+        if (expandedCard) {
+            HorizontalDivider(
+                thickness = 2.dp,
+                color = MaterialTheme.colorScheme.onPrimaryContainer,
+            )
+            Spacer(modifier = Modifier.height(7.dp))
+
             Row(
-                modifier =
-                    Modifier.fillMaxHeight(),
-                verticalAlignment = Alignment.CenterVertically,
+                modifier = modifier.fillMaxWidth(),
+                horizontalArrangement = Arrangement.SpaceBetween,
             ) {
-                if (expandedCard) {
+                Column(
+                    modifier = Modifier.fillMaxWidth(0.8f),
+                ) {
+                    DeviceInfo(device = deviceOverview.device)
+
+                    Spacer(modifier = Modifier.height(7.dp))
+
+                    DeviceStatus(
+                        deviceOverview = deviceOverview,
+                        state = deviceState,
+                    )
+                }
+
+                Column(
+                    modifier = Modifier.fillMaxWidth(0.5f),
+                    horizontalAlignment = Alignment.CenterHorizontally,
+                ) {
                     DeleteDeviceIconButton(
                         device = deviceOverview.device,
                         onRemove = {
@@ -154,31 +187,7 @@ fun DeviceCard(
                         },
                     )
                 }
-
-                Icon(
-                    modifier = Modifier.rotate(rotationState),
-                    imageVector = Icons.Default.ArrowDropDown,
-                    contentDescription = "Drop-Down Arrow",
-                )
             }
-        }
-        if (expandedCard) {
-            HorizontalDivider(
-                thickness = 2.dp,
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-            )
-
-            Spacer(modifier = Modifier.height(7.dp))
-
-            DeviceInfo(modifier, deviceOverview.device)
-
-            Spacer(modifier = Modifier.height(7.dp))
-
-            DeviceStatus(
-                modifier = modifier,
-                deviceOverview = deviceOverview,
-                state = deviceState,
-            )
 
             Spacer(modifier = Modifier.height(7.dp))
         }
@@ -189,27 +198,20 @@ fun DeviceCard(
 fun DeviceStateIcon(state: DeviceState) {
     when (state) {
         DeviceState.Active -> {
-            Icon(
-                imageVector = Icons.Default.PowerSettingsNew,
-                contentDescription = "Active",
-                tint = MaterialTheme.colorScheme.success,
-            )
+            Circle(MaterialTheme.colorScheme.success)
         }
         DeviceState.Scheduled -> {
-            Icon(
-                imageVector = Icons.Default.Schedule,
-                contentDescription = "Scheduled",
-                tint = MaterialTheme.colorScheme.scheduled,
-            )
+            Circle(MaterialTheme.colorScheme.scheduled)
         }
         DeviceState.Inactive -> {
-            Icon(
-                imageVector = Icons.Default.PowerSettingsNew,
-                contentDescription = "Scheduled",
-                tint = MaterialTheme.colorScheme.error,
-            )
+            Circle(MaterialTheme.colorScheme.error)
         }
     }
+}
+
+@Composable
+fun Circle(color: Color) {
+    Box(modifier = Modifier.size(10.dp).clip(CircleShape).background(color))
 }
 
 @Composable
@@ -309,7 +311,6 @@ fun DeleteDeviceIconButton(
     }
 
     IconButton(
-        modifier = Modifier,
         onClick = {
             openConfirmDialog = true
         },
