+++ frontend/app/src/main/java/dk/scheduling/schedulingfrontend/device/DeviceState.kt
@@ -10,16 +10,14 @@ enum class DeviceState {
 }
 
 fun getDeviceState(deviceOverview: DeviceOverview): DeviceState {
+    val dateTimeNow = LocalDateTime.now()
     val event = deviceOverview.event ?: return DeviceState.Inactive
 
-    if (event.startTime.isAfter(LocalDateTime.now())) {
+    if (event.startTime.isAfter(dateTimeNow)) {
         return DeviceState.Scheduled
     }
 
-    if (event.startTime.isBefore(
-            LocalDateTime.now(),
-        ) && LocalDateTime.now().isBefore(event.startTime.plus(event.duration, ChronoUnit.MILLIS))
-    ) {
+    if (event.startTime.isBefore(dateTimeNow) && dateTimeNow.isBefore(event.startTime.plus(event.duration, ChronoUnit.MILLIS))) {
         return DeviceState.Active
     }
